+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     
 -0.2.3
     - Now compatible with PHP 8 GD Image types
+    - Dataset cast sample to array upon validation
 
 - 0.2.2
     - Optimized CART quantile-based node splitting