+++ CHANGELOG.md
@@ -1,4 +1,4 @@
--0.2.3
+- 0.2.3
     - Now compatible with PHP 8 GD Image types
     - Dataset cast sample to array upon validation